Door and Ball
Prefab and materials I've been using for the door. I have two doors. I didn't get to fixing it up so it only works on the new input system just yet.

Same with the ball. However, you can hold it and let it go, provided it's not on the floor.